Factor out determination of plot bounds and delta
