--- crates/re_space_view_time_series/src/legacy_visualizer_system.rs
@@ -1,1 +1,0 @@
-
